Open Virtualization driver
Sierraware's Open Virtualization driver. It allows to communicate to Open
Virtualization's TEE to execute sensitive tasks in the secure world.

This patch is based on SierraWare's Open Virtualization kernel driver and header
files. This code is not directly usable in kernel space; the intention behind
this patch is to give Sierraware the recognition of creating the base driver to
communicate to Open Virtualization.

Copyright and contribution belongs to Sierraware. Documentation can be found at
www.openvirtualization.org

Signed-off-by: Javier Gonzalez <jgon@itu.dk>